+++ src/main/java/ch/supsi/rubattu/distance/EuclideanDistance.java
@@ -4,9 +4,9 @@ import ch.supsi.rubattu.model.City;
 
 public class EuclideanDistance implements Distance {
 
-    public double calculate(City c1, City c2) {
+    public int calculate(City c1, City c2) {
         double x = c1.x() - c2.x();
         double y = c1.y() - c2.y();
-        return Math.round(Math.sqrt(x*x + y*y));
+        return (int) Math.round(Math.sqrt(x*x + y*y));
     }
 }
